exclusão da classe de validação
